Incorporated feedback points from evaluation

+++ app/controllers/likes_controller.rb
@@ -3,18 +3,24 @@
 # Like controller
 class LikesController < ApplicationController
   def create
-    @like = current_user.likes.create(like_params)
-    @post = @like.post
-    respond_to :js
+    @like = current_user.likes.build(like_params)
+    authorize @like
+    if @like.save
+      @post = @like.post
+      respond_to :js
+    else
+      redirect_to(post_path, alert: t('.alert'))
+    end
   end
 
   def destroy
     @like = Like.find(params[:id])
     @post = @like.post
+    authorize @like
     if @like.destroy
       respond_to :js
     else
-      flash[:alert] = 'Something went wrong'
+      flash[:alert] = t('.alert')
     end
   end
